Switch to 2-week periods and support fractional targets
- Change tracking period from 1 week to 2 weeks (14 days)
- Use consistent 2-week boundaries anchored to a fixed reference date
- Support fractional weekly_target values (e.g., 1.5 for "3 per 2 weeks")
- Integer targets show segmented progress bars
- Fractional targets show smooth progress bars without segments
- Header now shows "Day X of 14" and 2-week date range
- Target marker still moves smoothly based on time of day

+++ app/ha/models.py
@@ -8,7 +8,7 @@ from typing import Optional
 class GoalConfig:
     """Configuration parsed from HA label."""
     label_id: str
-    weekly_target: int
+    weekly_target: float  # Target for the 2-week period (supports fractional, e.g., 1.5)
     emoji: Optional[str] = None
     sound: Optional[str] = None
